split linkTo into files
