+++ src/libs/Navigation/linkTo/index.ts
@@ -1,120 +1,32 @@
 import {getActionFromState} from '@react-navigation/core';
 import {findFocusedRoute} from '@react-navigation/native';
-import type {NavigationAction, NavigationContainerRef, NavigationState, PartialState} from '@react-navigation/native';
+import type {NavigationContainerRef, NavigationState, PartialState} from '@react-navigation/native';
 import {omitBy} from 'lodash';
-import type {Writable} from 'type-fest';
 import getIsNarrowLayout from '@libs/getIsNarrowLayout';
 import shallowCompare from '@libs/ObjectUtils';
 import {extractPolicyIDFromPath, getPathWithoutPolicyID} from '@libs/PolicyUtils';
+import getActionsFromPartialDiff from '@navigation/AppNavigator/getActionsFromPartialDiff';
+import getPartialStateDiff from '@navigation/AppNavigator/getPartialStateDiff';
+import dismissModal from '@navigation/dismissModal';
+import extrapolateStateFromParams from '@navigation/extrapolateStateFromParams';
+import getPolicyIDFromState from '@navigation/getPolicyIDFromState';
+import getStateFromPath from '@navigation/getStateFromPath';
+import getTopmostBottomTabRoute from '@navigation/getTopmostBottomTabRoute';
+import getTopmostCentralPaneRoute from '@navigation/getTopmostCentralPaneRoute';
+import getTopmostReportId from '@navigation/getTopmostReportId';
+import isSideModalNavigator from '@navigation/isSideModalNavigator';
+import linkingConfig from '@navigation/linkingConfig';
+import getAdaptedStateFromPath from '@navigation/linkingConfig/getAdaptedStateFromPath';
+import getMatchingBottomTabRouteForState from '@navigation/linkingConfig/getMatchingBottomTabRouteForState';
+import getMatchingCentralPaneRouteForState from '@navigation/linkingConfig/getMatchingCentralPaneRouteForState';
+import replacePathInNestedState from '@navigation/linkingConfig/replacePathInNestedState';
+import type {NavigationRoot, RootStackParamList, StackNavigationAction, State} from '@navigation/types';
 import CONST from '@src/CONST';
 import NAVIGATORS from '@src/NAVIGATORS';
 import type {Route} from '@src/ROUTES';
 import SCREENS from '@src/SCREENS';
-import getActionsFromPartialDiff from './AppNavigator/getActionsFromPartialDiff';
-import getPartialStateDiff from './AppNavigator/getPartialStateDiff';
-import dismissModal from './dismissModal';
-import extrapolateStateFromParams from './extrapolateStateFromParams';
-import getPolicyIDFromState from './getPolicyIDFromState';
-import getStateFromPath from './getStateFromPath';
-import getTopmostBottomTabRoute from './getTopmostBottomTabRoute';
-import getTopmostCentralPaneRoute from './getTopmostCentralPaneRoute';
-import getTopmostReportId from './getTopmostReportId';
-import isSideModalNavigator from './isSideModalNavigator';
-import linkingConfig from './linkingConfig';
-import getAdaptedStateFromPath from './linkingConfig/getAdaptedStateFromPath';
-import getMatchingBottomTabRouteForState from './linkingConfig/getMatchingBottomTabRouteForState';
-import getMatchingCentralPaneRouteForState from './linkingConfig/getMatchingCentralPaneRouteForState';
-import replacePathInNestedState from './linkingConfig/replacePathInNestedState';
-import type {NavigationRoot, RootStackParamList, StackNavigationAction, State} from './types';
-
-type ActionPayloadParams = {
-    screen?: string;
-    params?: unknown;
-    path?: string;
-};
-
-type ActionPayload = {
-    params?: ActionPayloadParams;
-};
-
-/**
- * Motivation for this function is described in NAVIGATION.md
- *
- * @param action action generated by getActionFromState
- * @param state The root state
- * @returns minimalAction minimal action is the action that we should dispatch
- */
-function getMinimalAction(action: NavigationAction, state: NavigationState): Writable<NavigationAction> {
-    let currentAction: NavigationAction = action;
-    let currentState: State | undefined = state;
-    let currentTargetKey: string | undefined;
-
-    while (currentAction.payload && 'name' in currentAction.payload && currentState?.routes[currentState.index ?? -1].name === currentAction.payload.name) {
-        if (!currentState?.routes[currentState.index ?? -1].state) {
-            break;
-        }
-
-        currentState = currentState?.routes[currentState.index ?? -1].state;
-        currentTargetKey = currentState?.key;
-
-        const payload = currentAction.payload as ActionPayload;
-
-        // Creating new smaller action
-        currentAction = {
-            type: currentAction.type,
-            payload: {
-                name: payload?.params?.screen,
-                params: payload?.params?.params,
-                path: payload?.params?.path,
-            },
-            target: currentTargetKey,
-        };
-    }
-    return currentAction;
-}
-
-// Because we need to change the type to push, we also need to set target for this action to the bottom tab navigator.
-function getActionForBottomTabNavigator(
-    action: StackNavigationAction,
-    state: NavigationState<RootStackParamList>,
-    policyID?: string,
-    shouldNavigate?: boolean,
-): Writable<NavigationAction> | undefined {
-    const bottomTabNavigatorRoute = state.routes.at(0);
-
-    if (!bottomTabNavigatorRoute || bottomTabNavigatorRoute.state === undefined || !action || action.type !== CONST.NAVIGATION.ACTION_TYPE.NAVIGATE) {
-        return;
-    }
-
-    const params = action.payload.params as ActionPayloadParams;
-    let payloadParams = params.params as Record<string, string | undefined>;
-    const screen = params.screen;
-
-    if (!payloadParams) {
-        payloadParams = {policyID};
-    } else if (!('policyID' in payloadParams && !!payloadParams?.policyID)) {
-        payloadParams = {...payloadParams, policyID};
-    }
-
-    // Check if the current bottom tab is the same as the one we want to navigate to. If it is, we don't need to do anything.
-    const bottomTabCurrentTab = getTopmostBottomTabRoute(state);
-    const bottomTabParams = bottomTabCurrentTab?.params as Record<string, string | undefined>;
-
-    // Verify if the policyID is different than the one we are currently on. If it is, we need to navigate to the new policyID.
-    const isNewPolicy = bottomTabParams?.policyID !== payloadParams?.policyID;
-    if (bottomTabCurrentTab?.name === screen && !shouldNavigate && !isNewPolicy) {
-        return;
-    }
-
-    return {
-        type: CONST.NAVIGATION.ACTION_TYPE.PUSH,
-        payload: {
-            name: screen,
-            params: payloadParams,
-        },
-        target: bottomTabNavigatorRoute.state.key,
-    };
-}
+import getActionForBottomTabNavigator from './getActionForBottomTabNavigator';
+import getMinimalAction from './getMinimalAction';
 
 export default function linkTo(navigation: NavigationContainerRef<RootStackParamList> | null, path: Route, type?: string, isActiveRoute?: boolean) {
     if (!navigation) {
@@ -127,6 +39,7 @@ export default function linkTo(navigation: NavigationContainerRef<RootStackParam
     while ((current = root.getParent())) {
         root = current;
     }
+
     const pathWithoutPolicyID = getPathWithoutPolicyID(`/${path}`) as Route;
     const rootState = navigation.getRootState() as NavigationState<RootStackParamList>;
     const stateFromPath = getStateFromPath(pathWithoutPolicyID) as PartialState<NavigationState<RootStackParamList>>;
@@ -258,6 +171,7 @@ export default function linkTo(navigation: NavigationContainerRef<RootStackParam
     }
 
     if (action && 'payload' in action && action.payload && 'name' in action.payload && isSideModalNavigator(action.payload.name)) {
+        // Information about the state may be in the params.
         const currentFocusedRoute = findFocusedRoute(extrapolateStateFromParams(rootState));
         const targetFocusedRoute = findFocusedRoute(stateFromPath);
